reorganized pages folder to have subfolder per page
