+++ client/src/pages/static/About.tsx
