+++ client/src/pages/static/Home.tsx
