+++ client/src/pages/static/Product.tsx
