+++ client/src/pages/static/index.ts
@@ -1,0 +1,5 @@
+import About from './About';
+import Home from './Home';
+import Product from './Product';
+
+export { About, Home, Product };
