+++ client/src/pages/upload/index.ts
@@ -1,0 +1,3 @@
+import Upload from './Upload';
+
+export default Upload;
